#3 Renamed git.py to versioning.py, to avoid name clash with the corresponding test_git.py tests.
